+++ presto-raptor/src/main/java/com/facebook/presto/raptor/storage/OrcStorageManager.java
@@ -14,17 +14,12 @@
 package com.facebook.presto.raptor.storage;
 
 import com.facebook.airlift.json.JsonCodec;
-import com.facebook.presto.memory.context.AggregatedMemoryContext;
-import com.facebook.presto.orc.OrcBatchRecordReader;
 import com.facebook.presto.orc.OrcDataSink;
 import com.facebook.presto.orc.OrcDataSource;
-import com.facebook.presto.orc.OrcPredicate;
 import com.facebook.presto.orc.OrcReader;
 import com.facebook.presto.orc.OrcReaderOptions;
 import com.facebook.presto.orc.OrcWriterStats;
 import com.facebook.presto.orc.StripeMetadataSource;
-import com.facebook.presto.orc.TupleDomainOrcPredicate;
-import com.facebook.presto.orc.TupleDomainOrcPredicate.ColumnReference;
 import com.facebook.presto.orc.cache.OrcFileTailSource;
 import com.facebook.presto.orc.metadata.CompressionKind;
 import com.facebook.presto.orc.metadata.OrcType;
@@ -44,21 +39,10 @@ import com.facebook.presto.spi.NodeManager;
 import com.facebook.presto.spi.Page;
 import com.facebook.presto.spi.PrestoException;
 import com.facebook.presto.spi.predicate.TupleDomain;
-import com.facebook.presto.spi.type.ArrayType;
-import com.facebook.presto.spi.type.DecimalType;
-import com.facebook.presto.spi.type.MapType;
-import com.facebook.presto.spi.type.NamedTypeSignature;
-import com.facebook.presto.spi.type.RowFieldName;
-import com.facebook.presto.spi.type.RowType;
-import com.facebook.presto.spi.type.StandardTypes;
-import com.facebook.presto.spi.type.TimestampType;
 import com.facebook.presto.spi.type.Type;
 import com.facebook.presto.spi.type.TypeManager;
-import com.facebook.presto.spi.type.TypeSignature;
-import com.facebook.presto.spi.type.TypeSignatureParameter;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
 import io.airlift.slice.Slice;
 import io.airlift.slice.Slices;
@@ -66,12 +50,10 @@ import io.airlift.units.DataSize;
 import io.airlift.units.Duration;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
-import org.joda.time.DateTimeZone;
 
 import javax.annotation.PreDestroy;
 import javax.inject.Inject;
 
-import java.io.Closeable;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.BitSet;
@@ -94,36 +76,20 @@ import static com.facebook.airlift.concurrent.MoreFutures.allAsList;
 import static com.facebook.airlift.concurrent.MoreFutures.getFutureValue;
 import static com.facebook.airlift.concurrent.Threads.daemonThreadsNamed;
 import static com.facebook.airlift.json.JsonCodec.jsonCodec;
-import static com.facebook.presto.memory.context.AggregatedMemoryContext.newSimpleAggregatedMemoryContext;
 import static com.facebook.presto.orc.OrcEncoding.ORC;
-import static com.facebook.presto.orc.OrcReader.INITIAL_BATCH_SIZE;
-import static com.facebook.presto.raptor.RaptorColumnHandle.isBucketNumberColumn;
-import static com.facebook.presto.raptor.RaptorColumnHandle.isHiddenColumn;
-import static com.facebook.presto.raptor.RaptorColumnHandle.isShardRowIdColumn;
-import static com.facebook.presto.raptor.RaptorColumnHandle.isShardUuidColumn;
 import static com.facebook.presto.raptor.RaptorErrorCode.RAPTOR_ERROR;
 import static com.facebook.presto.raptor.RaptorErrorCode.RAPTOR_LOCAL_DISK_FULL;
 import static com.facebook.presto.raptor.RaptorErrorCode.RAPTOR_RECOVERY_ERROR;
 import static com.facebook.presto.raptor.RaptorErrorCode.RAPTOR_RECOVERY_TIMEOUT;
 import static com.facebook.presto.raptor.filesystem.FileSystemUtil.xxhash64;
-import static com.facebook.presto.raptor.storage.OrcPageSource.BUCKET_NUMBER_COLUMN;
-import static com.facebook.presto.raptor.storage.OrcPageSource.NULL_COLUMN;
-import static com.facebook.presto.raptor.storage.OrcPageSource.ROWID_COLUMN;
-import static com.facebook.presto.raptor.storage.OrcPageSource.SHARD_UUID_COLUMN;
 import static com.facebook.presto.raptor.storage.ShardStats.computeColumnStats;
 import static com.facebook.presto.raptor.storage.StorageManagerConfig.OrcOptimizedWriterStage.ENABLED_AND_VALIDATED;
-import static com.facebook.presto.spi.type.BigintType.BIGINT;
-import static com.facebook.presto.spi.type.BooleanType.BOOLEAN;
-import static com.facebook.presto.spi.type.CharType.createCharType;
-import static com.facebook.presto.spi.type.DoubleType.DOUBLE;
-import static com.facebook.presto.spi.type.VarbinaryType.VARBINARY;
-import static com.facebook.presto.spi.type.VarcharType.createUnboundedVarcharType;
-import static com.facebook.presto.spi.type.VarcharType.createVarcharType;
+import static com.facebook.presto.raptor.storage.StorageManagerUtils.HUGE_MAX_READ_BLOCK_SIZE;
+import static com.facebook.presto.raptor.storage.StorageManagerUtils.closeQuietly;
+import static com.facebook.presto.raptor.storage.StorageManagerUtils.getType;
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkState;
 import static com.google.common.base.Throwables.throwIfInstanceOf;
-import static com.google.common.collect.ImmutableList.toImmutableList;
-import static io.airlift.units.DataSize.Unit.PETABYTE;
 import static java.lang.Math.min;
 import static java.lang.String.format;
 import static java.util.Objects.requireNonNull;
@@ -132,19 +98,10 @@ import static java.util.concurrent.CompletableFuture.supplyAsync;
 import static java.util.concurrent.Executors.newCachedThreadPool;
 import static java.util.concurrent.Executors.newFixedThreadPool;
 import static java.util.stream.Collectors.toList;
-import static org.joda.time.DateTimeZone.UTC;
 
 public class OrcStorageManager
         implements StorageManager
 {
-    // Raptor does not store time-related data types as they are in ORC.
-    // They will be casted to BIGINT or INTEGER to avoid timezone conversion.
-    // This is due to the historical reason of using the legacy ORC read/writer that does not support timestamp types.
-    // In order to be consistent, we still enforce the conversion.
-    // The following DEFAULT_STORAGE_TIMEZONE is not used by the optimized ORC read/writer given we never read/write timestamp types.
-    public static final DateTimeZone DEFAULT_STORAGE_TIMEZONE = UTC;
-    // TODO: do not limit the max size of blocks to read for now; enable the limit when the Hive connector is ready
-    public static final DataSize HUGE_MAX_READ_BLOCK_SIZE = new DataSize(1, PETABYTE);
     private static final JsonCodec<ShardDelta> SHARD_DELTA_CODEC = jsonCodec(ShardDelta.class);
 
     private static final long MAX_ROWS = 1_000_000_000;
@@ -285,8 +242,6 @@ public class OrcStorageManager
         FileSystem fileSystem = orcDataEnvironment.getFileSystem(fileSystemContext);
         OrcDataSource dataSource = openShard(fileSystem, shardUuid, readerAttributes);
 
-        AggregatedMemoryContext systemMemoryUsage = newSimpleAggregatedMemoryContext();
-
         try {
             OrcReader reader = new OrcReader(
                     dataSource,
@@ -295,62 +250,27 @@ public class OrcStorageManager
                     stripeMetadataSource,
                     new OrcReaderOptions(readerAttributes.getMaxMergeDistance(), readerAttributes.getTinyStripeThreshold(), HUGE_MAX_READ_BLOCK_SIZE, readerAttributes.isZstdJniDecompressionEnabled()));
 
-            Map<Long, Integer> indexMap = columnIdIndex(reader.getColumnNames());
-            ImmutableMap.Builder<Integer, Type> includedColumns = ImmutableMap.builder();
-            ImmutableList.Builder<Integer> columnIndexes = ImmutableList.builder();
-            for (int i = 0; i < columnIds.size(); i++) {
-                long columnId = columnIds.get(i);
-                if (isHiddenColumn(columnId)) {
-                    columnIndexes.add(toSpecialIndex(columnId));
-                    continue;
-                }
-
-                Integer index = indexMap.get(columnId);
-                if (index == null) {
-                    columnIndexes.add(NULL_COLUMN);
-                }
-                else {
-                    columnIndexes.add(index);
-                    includedColumns.put(index, toOrcFileType(columnTypes.get(i), typeManager));
-                }
-            }
-
-            OrcPredicate predicate = getPredicate(effectivePredicate, indexMap);
-
-            StorageTypeConverter storageTypeConverter = new StorageTypeConverter(typeManager);
-
-            OrcBatchRecordReader recordReader = reader.createBatchRecordReader(storageTypeConverter.toStorageTypes(includedColumns.build()), predicate, DEFAULT_STORAGE_TIMEZONE, systemMemoryUsage, INITIAL_BATCH_SIZE);
-
             Optional<ShardRewriter> shardRewriter = Optional.empty();
             if (transactionId.isPresent()) {
                 checkState(allColumnTypes.isPresent());
                 shardRewriter = Optional.of(createShardRewriter(fileSystem, transactionId.getAsLong(), bucketNumber, shardUuid, allColumnTypes.get()));
             }
 
-            return new OrcPageSource(shardRewriter, recordReader, dataSource, columnIds, columnTypes, columnIndexes.build(), shardUuid, bucketNumber, systemMemoryUsage);
+            return StorageManagerUtils.getPageSource(
+                    reader,
+                    dataSource,
+                    typeManager,
+                    shardUuid,
+                    bucketNumber,
+                    columnIds,
+                    columnTypes,
+                    effectivePredicate,
+                    shardRewriter);
         }
         catch (IOException | RuntimeException e) {
             closeQuietly(dataSource);
             throw new PrestoException(RAPTOR_ERROR, "Failed to create page source for shard " + shardUuid, e);
         }
-        catch (Throwable t) {
-            closeQuietly(dataSource);
-            throw t;
-        }
-    }
-
-    private static int toSpecialIndex(long columnId)
-    {
-        if (isShardRowIdColumn(columnId)) {
-            return ROWID_COLUMN;
-        }
-        if (isShardUuidColumn(columnId)) {
-            return SHARD_UUID_COLUMN;
-        }
-        if (isBucketNumberColumn(columnId)) {
-            return BUCKET_NUMBER_COLUMN;
-        }
-        throw new PrestoException(RAPTOR_ERROR, "Invalid column ID: " + columnId);
     }
 
     @Override
@@ -522,7 +442,7 @@ public class OrcStorageManager
 
     private List<ColumnInfo> getColumnInfoFromOrcColumnTypes(List<String> orcColumnNames, List<OrcType> orcColumnTypes)
     {
-        Type rowType = getType(orcColumnTypes, 0);
+        Type rowType = getType(typeManager, orcColumnTypes, 0);
         if (orcColumnNames.size() != rowType.getTypeParameters().size()) {
             throw new PrestoException(RAPTOR_ERROR, "Column names and types do not match");
         }
@@ -547,91 +467,6 @@ public class OrcStorageManager
                 .sorted(Map.Entry.comparingByKey())
                 .map(entry -> new ColumnInfo(entry.getKey(), typeManager.getType(entry.getValue())))
                 .collect(toList());
-    }
-
-    private Type getType(List<OrcType> types, int index)
-    {
-        OrcType type = types.get(index);
-        switch (type.getOrcTypeKind()) {
-            case BOOLEAN:
-                return BOOLEAN;
-            case LONG:
-                return BIGINT;
-            case DOUBLE:
-                return DOUBLE;
-            case STRING:
-                return createUnboundedVarcharType();
-            case VARCHAR:
-                return createVarcharType(type.getLength().get());
-            case CHAR:
-                return createCharType(type.getLength().get());
-            case BINARY:
-                return VARBINARY;
-            case DECIMAL:
-                return DecimalType.createDecimalType(type.getPrecision().get(), type.getScale().get());
-            case LIST:
-                TypeSignature elementType = getType(types, type.getFieldTypeIndex(0)).getTypeSignature();
-                return typeManager.getParameterizedType(StandardTypes.ARRAY, ImmutableList.of(TypeSignatureParameter.of(elementType)));
-            case MAP:
-                TypeSignature keyType = getType(types, type.getFieldTypeIndex(0)).getTypeSignature();
-                TypeSignature valueType = getType(types, type.getFieldTypeIndex(1)).getTypeSignature();
-                return typeManager.getParameterizedType(StandardTypes.MAP, ImmutableList.of(TypeSignatureParameter.of(keyType), TypeSignatureParameter.of(valueType)));
-            case STRUCT:
-                List<String> fieldNames = type.getFieldNames();
-                ImmutableList.Builder<TypeSignatureParameter> fieldTypes = ImmutableList.builder();
-                for (int i = 0; i < type.getFieldCount(); i++) {
-                    fieldTypes.add(TypeSignatureParameter.of(new NamedTypeSignature(
-                            Optional.of(new RowFieldName(fieldNames.get(i), false)),
-                            getType(types, type.getFieldTypeIndex(i)).getTypeSignature())));
-                }
-                return typeManager.getParameterizedType(StandardTypes.ROW, fieldTypes.build());
-        }
-        throw new PrestoException(RAPTOR_ERROR, "Unhandled ORC type: " + type);
-    }
-
-    static Type toOrcFileType(Type raptorType, TypeManager typeManager)
-    {
-        // TIMESTAMPS are stored as BIGINT to void the poor encoding in ORC
-        if (raptorType == TimestampType.TIMESTAMP) {
-            return BIGINT;
-        }
-        if (raptorType instanceof ArrayType) {
-            Type elementType = toOrcFileType(((ArrayType) raptorType).getElementType(), typeManager);
-            return new ArrayType(elementType);
-        }
-        if (raptorType instanceof MapType) {
-            TypeSignature keyType = toOrcFileType(((MapType) raptorType).getKeyType(), typeManager).getTypeSignature();
-            TypeSignature valueType = toOrcFileType(((MapType) raptorType).getValueType(), typeManager).getTypeSignature();
-            return typeManager.getParameterizedType(StandardTypes.MAP, ImmutableList.of(TypeSignatureParameter.of(keyType), TypeSignatureParameter.of(valueType)));
-        }
-        if (raptorType instanceof RowType) {
-            List<RowType.Field> fields = ((RowType) raptorType).getFields().stream()
-                    .map(field -> new RowType.Field(field.getName(), toOrcFileType(field.getType(), typeManager)))
-                    .collect(toImmutableList());
-            return RowType.from(fields);
-        }
-        return raptorType;
-    }
-
-    private static OrcPredicate getPredicate(TupleDomain<RaptorColumnHandle> effectivePredicate, Map<Long, Integer> indexMap)
-    {
-        ImmutableList.Builder<ColumnReference<RaptorColumnHandle>> columns = ImmutableList.builder();
-        for (RaptorColumnHandle column : effectivePredicate.getDomains().get().keySet()) {
-            Integer index = indexMap.get(column.getColumnId());
-            if (index != null) {
-                columns.add(new ColumnReference<>(column, index, column.getColumnType()));
-            }
-        }
-        return new TupleDomainOrcPredicate<>(effectivePredicate, columns.build(), false, Optional.empty());
-    }
-
-    private static Map<Long, Integer> columnIdIndex(List<String> columnNames)
-    {
-        ImmutableMap.Builder<Long, Integer> map = ImmutableMap.builder();
-        for (int i = 0; i < columnNames.size(); i++) {
-            map.put(Long.valueOf(columnNames.get(i)), i);
-        }
-        return map.build();
     }
 
     private class OrcStoragePageSink
@@ -786,15 +621,6 @@ public class OrcStorageManager
                 }
                 writer = new OrcFileWriter(columnIds, columnTypes, sink, orcOptimizedWriterStage.equals(ENABLED_AND_VALIDATED), stats, typeManager, compression);
             }
-        }
-    }
-
-    private static void closeQuietly(Closeable closeable)
-    {
-        try {
-            closeable.close();
-        }
-        catch (IOException ignored) {
         }
     }
 }
